Bolt on check raposo-pulido kepler algo.

+++ harmonica/orbit/trajectories.hpp
@@ -5,7 +5,7 @@
 #include <pybind11/pybind11.h>
 
 
-void orbital_trajectories(int i, int j,
+void orbital_trajectories(double i, double j,
                           bool require_gradients,
                           pybind11::array_t<double, pybind11::array::c_style> np_in,
                           pybind11::array_t<double, pybind11::array::c_style> np_out);
